extract redundant code in custom hook
