+++ apps/frontend/src/view/pages/risk.page.tsx
@@ -10,16 +10,7 @@ import {
     TableSortLabel,
     Typography,
 } from "@mui/material";
-import {
-    memo,
-    useEffect,
-    useLayoutEffect,
-    useRef,
-    useState,
-    type ChangeEvent,
-    type ReactNode,
-    type SyntheticEvent,
-} from "react";
+import { memo, useLayoutEffect, useState, type ChangeEvent, type ReactNode, type SyntheticEvent } from "react";
 
 import { useTranslation } from "react-i18next";
 import { Route, Routes, useNavigate, useParams } from "react-router";
@@ -33,6 +24,7 @@ import { checkUserRole, USER_ROLES } from "#api/types/user-roles.types.ts";
 import { NavigationActions } from "#application/actions/navigation.actions.ts";
 import { ProjectsActions } from "#application/actions/projects.actions.ts";
 import { useConfirm } from "#application/hooks/use-confirm.hook.ts";
+import { useLoadThreatsOnce } from "#application/hooks/use-load-threats-once.hook.ts";
 import { useMatrix } from "#application/hooks/use-matrix.hook.ts";
 import { MATRIX_COLOR } from "#view/colors/matrix.ts";
 import { IconButton } from "#view/components/icon-button.component.tsx";
@@ -263,21 +255,7 @@ const RiskPageBody = ({ project }: RiskPageBodyProps) => {
         }
     };
 
-    const prevProjectIdRef = useRef(projectId);
-    const loadedRef = useRef(false);
-    useEffect(() => {
-        if (prevProjectIdRef.current !== projectId) {
-            prevProjectIdRef.current = projectId;
-            loadedRef.current = false;
-        }
-        if (loadedRef.current) {
-            return;
-        }
-        if (autoSaveStatus === "upToDate" || autoSaveStatus === "uninitialized") {
-            loadedRef.current = true;
-            loadThreats();
-        }
-    }, [projectId, autoSaveStatus, loadThreats]);
+    useLoadThreatsOnce({ projectId, autoSaveStatus, load: loadThreats });
 
     return (
         <Page sx={{ boxSizing: "border-box" }}>
