+++ apps/frontend/src/view/pages/threats.page.tsx
@@ -6,7 +6,7 @@ import TableCell, { type TableCellProps } from "@mui/material/TableCell";
 import TableContainer from "@mui/material/TableContainer";
 import TableHead from "@mui/material/TableHead";
 import TableRow from "@mui/material/TableRow";
-import { memo, useEffect, useLayoutEffect, useRef, useState, type ChangeEvent, type SyntheticEvent } from "react";
+import { memo, useLayoutEffect, useState, type ChangeEvent, type SyntheticEvent } from "react";
 import { useTranslation } from "react-i18next";
 import { Route, Routes, useNavigate, useParams } from "react-router";
 import type { ExtendedThreat } from "#api/types/threat.types.ts";
@@ -14,6 +14,7 @@ import { checkUserRole, USER_ROLES } from "#api/types/user-roles.types.ts";
 import { NavigationActions } from "#application/actions/navigation.actions.ts";
 import { useConfirm } from "#application/hooks/use-confirm.hook.ts";
 import { useEditor } from "#application/hooks/use-editor.hook.ts";
+import { useLoadThreatsOnce } from "#application/hooks/use-load-threats-once.hook.ts";
 import { useThreatsList, type ThreatListItem } from "#application/hooks/use-threats-list.hook.ts";
 import { IconButton } from "#view/components/icon-button.component.tsx";
 import { Page } from "#view/components/page.component.tsx";
@@ -124,21 +125,7 @@ const ThreatsPageBody = () => {
         });
     };
 
-    const prevProjectIdRef = useRef(projectId);
-    const loadedRef = useRef(false);
-    useEffect(() => {
-        if (prevProjectIdRef.current !== projectId) {
-            prevProjectIdRef.current = projectId;
-            loadedRef.current = false;
-        }
-        if (loadedRef.current) {
-            return;
-        }
-        if (autoSaveStatus === "upToDate" || autoSaveStatus === "uninitialized") {
-            loadedRef.current = true;
-            loadThreats();
-        }
-    }, [projectId, autoSaveStatus, loadThreats]);
+    useLoadThreatsOnce({ projectId, autoSaveStatus, load: loadThreats });
 
     const [assetAnchorEl, setAssetAnchorEl] = useState<HTMLElement | null>(null);
     const [currentAssetList, setCurrentAssetList] = useState<ExtendedThreat["assets"] | null>(null);
